[FIX] Mapper V2 테스트 완료
