+++ server/src/main/java/com/main026/walking/pet/mapper/PetMapper.java
@@ -1,4 +1,3 @@
-/*
 package com.main026.walking.pet.mapper;
 
 import com.main026.walking.community.dto.CommunityDto;
@@ -6,16 +5,22 @@ import com.main026.walking.community.entity.Community;
 import com.main026.walking.member.dto.MemberDto;
 import com.main026.walking.pet.dto.PetDto;
 import com.main026.walking.pet.entity.Pet;
+import com.main026.walking.util.awsS3.AwsS3Service;
 import org.mapstruct.Mapper;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Mapper(componentModel = "spring")
-public interface PetMapper {
+public abstract class PetMapper {
 
-    Pet petPostDtoToPet(PetDto.Post postDto);
-    default PetDto.Response petToPetResponseDto(Pet pet){
+    @Autowired
+    AwsS3Service awsS3Service;
+
+    public abstract Pet petPostDtoToPet(PetDto.Post postDto);
+
+    public PetDto.Response petToPetResponseDto(Pet pet){
         if(pet == null){
             return null;
         }
@@ -26,17 +31,20 @@ public interface PetMapper {
         response.petAges(pet.getPetAges());
         response.neuter(pet.getNeuter());
         response.personality(pet.getPersonality());
-        response.imgUrl(pet.getImgUrl());
+        response.imgUrl(awsS3Service.getFileURL(pet.getImgUrl()));
         response.breed(pet.getBreed());
         response.about(pet.getAbout());
 
-        response.member(new MemberDto.Response(pet.getMember()));
+//        response.member(new MemberDto.Response(pet.getMember()));
+        MemberDto.compactResponse dto = new MemberDto.compactResponse(pet.getMember());
+        dto.setImgUrl(awsS3Service.getFileURL(dto.getImgUrl()));
+        response.member(dto);
 
         List<CommunityDto.compactResponse> petCommunityList = new ArrayList<>();
         for (int i = 0; i < pet.getCommunityPets().size(); i++) {
             Community community = pet.getCommunityPets().get(i).getCommunity();
             CommunityDto.compactResponse compactResponse = new CommunityDto.compactResponse(community);
-
+            compactResponse.setRepresentImgUrls(awsS3Service.getFileURL(compactResponse.getRepresentImgUrls()));
             if (!petCommunityList.contains(compactResponse)){
                 petCommunityList.add(compactResponse);
             }
@@ -47,4 +55,3 @@ public interface PetMapper {
     }
 
 }
-*/
